# [static runtime] Add NNC