--- a/torch/csrc/jit/passes/utils/subgraph_utils.cpp
+++ b/torch/csrc/jit/passes/utils/subgraph_utils.cpp
@@ -193,7 +193,8 @@ void mergeNodeIntoSubgraph(
     if (inputsMap.count(input) == 0) {
       // Clone constants inside the subgraph instead of referencing them, to
       // enable more optimizations
-      if (auto value = toIValue(input)) {
+      auto value = toIValue(input);
+      if (value && !value->isTensor()) {
         auto nv = subgraph->insertConstant(*value);
         nv->setType(input->type()); // Need to retain type information on Nones
         inputsMap[input] = nv;