--- a/torch/csrc/jit/runtime/static/impl.cpp
+++ b/torch/csrc/jit/runtime/static/impl.cpp
@@ -4,6 +4,7 @@
 #include <torch/csrc/jit/passes/freeze_module.h>
 #include <torch/csrc/jit/passes/remove_mutation.h>
 #include <torch/csrc/jit/passes/subgraph_rewrite.h>
+#include <torch/csrc/jit/passes/tensorexpr_fuser.h>
 #include <torch/csrc/jit/runtime/static/impl.h>
 #include <torch/csrc/jit/runtime/static/ops.h>
 #include <torch/csrc/jit/runtime/vararg_functions.h>
@@ -50,6 +51,8 @@ StaticRuntime::StaticRuntime(const torch::jit::Module& m)
       graph_->eraseInput(0);
     }
   }
+
+  FuseTensorExprs(graph_);
 
   // fill workspace_ with constants
   for (Node* node : graph_->nodes()) {
@@ -101,8 +104,10 @@ ProcessedNode::ProcessedNode(Node* node) : node_(node) {
   if (node->kind() != prim::ListConstruct &&
       node->kind() != prim::TupleConstruct) {
     const Operator& op = node->getOperator();
-    CHECK(op.hasOperation());
-    op_ = op.getOperation(node);
+    // static Symbol s = Symbol::fromQualString("tensorexpr::Group");
+    if (op.hasOperation()) {
+      op_ = op.getOperation(node);
+    }
   }
   if (canRunOutOfPlace(node)) {
     fn_ = getOutOfPlaceOperation(node);
@@ -110,48 +115,94 @@ ProcessedNode::ProcessedNode(Node* node) : node_(node) {
 }
 
 void ProcessedNode::run(StaticRuntime::ConstantMap& workspace) const {
-  if (!fn_) {
-    std::vector<IValue> stack;
-    const size_t size = node_->inputs().size();
-    stack.reserve(size);
-    for (size_t i = 0; i < size; i++) {
-      Value* v = node_->inputs()[i];
-      auto f = workspace.find(v);
-      TORCH_CHECK(
-          f != workspace.end(),
-          "Workspace does not contain Value ",
-          v->debugName());
-      stack.emplace_back(f->second);
-    }
-    if (op_) {
-      op_->operator()(&stack);
-    } else {
-      if (node_->kind() == prim::ListConstruct) {
-        listConstruct(
-            stack,
-            node_->output()->type()->expect<ListType>(),
-            node_->inputs().size());
-      } else if (node_->kind() == prim::TupleConstruct) {
-        bool named =
-            node_->output()->type()->expect<TupleType>()->name().has_value();
-        if (named) {
-          namedTupleConstruct(
-              stack,
-              node_->output()->type()->expect<TupleType>(),
-              node_->inputs().size());
-        } else {
-          tupleConstruct(stack, node_->inputs().size());
-        }
-      } else {
-        TORCH_CHECK(0, "Unhandled operation!", node_->kind().toQualString());
-      }
-    }
-    DCHECK_EQ(stack.size(), node_->outputs().size());
-    for (auto i = 0; i < node_->outputs().size(); i++) {
-      workspace[node_->outputs()[i]] = stack[i];
-    }
-  } else {
+  if (fn_) {
     fn_->operator()(workspace);
+    return;
+  }
+
+  std::vector<IValue> stack;
+  const size_t size = node_->inputs().size();
+  stack.reserve(size);
+
+  for (size_t i = 0; i < size; i++) {
+    Value* v = node_->inputs()[i];
+    auto f = workspace.find(v);
+    TORCH_CHECK(
+        f != workspace.end(),
+        "Workspace does not contain Value ",
+        v->debugName());
+    stack.emplace_back(f->second);
+  }
+
+  if (op_) {
+    op_->operator()(&stack);
+  } else {
+    if (node_->kind() == prim::ListConstruct) {
+      listConstruct(
+          stack,
+          node_->output()->type()->expect<ListType>(),
+          node_->inputs().size());
+    } else if (node_->kind() == prim::TupleConstruct) {
+      bool named =
+          node_->output()->type()->expect<TupleType>()->name().has_value();
+      if (named) {
+        namedTupleConstruct(
+            stack,
+            node_->output()->type()->expect<TupleType>(),
+            node_->inputs().size());
+      } else {
+        tupleConstruct(stack, node_->inputs().size());
+      }
+    } else if (node_->kind() == Symbol::fromQualString("tensorexpr::Group")) {
+      uint64_t h = 0;
+      auto hash = [](uint64_t x) {
+        x = ((x >> 16) ^ x) * 0x45d9f3b;
+        x = ((x >> 16) ^ x) * 0x45d9f3b;
+        x = (x >> 16) ^ x;
+        return x;
+      };
+      for (size_t i = 0; i < size; i++) {
+        if (stack[i].isTensor()) {
+          auto t = stack[i].toTensor();
+          h = hash(t.get_device() + h);
+          for (auto s : t.sizes()) {
+            h = hash(s + h);
+          }
+          for (auto s : t.strides()) {
+            h = hash(s + h);
+          }
+        } else if (stack[i].isInt()) {
+          h = hash(stack[i].toInt() + h);
+        }
+      }
+      // dispatch to precompiled if possible
+      if (!compiled_.count(h)) {
+        auto ops = torch::jit::getAllOperatorsFor(node_->kind());
+        CHECK(ops.size() == 1);
+        auto op = ops.front();
+        auto subgraph = node_->g(attr::Subgraph);
+        auto sg_is = subgraph->inputs();
+        auto node_is = node_->inputs();
+        CHECK(stack.size() == node_is.size());
+        CHECK(stack.size() == sg_is.size());
+        for (auto i = 0; i < stack.size(); ++i) {
+          if (stack[i].isTensor()) {
+            sg_is[i]->inferTypeFrom(stack[i].toTensor());
+            node_is[i]->inferTypeFrom(stack[i].toTensor());
+          }
+        }
+        auto operation = op->getOperation(node_);
+        CHECK(operation);
+        compiled_.emplace(h, operation);
+      }
+      compiled_.at(h)(&stack);
+    } else {
+      TORCH_CHECK(0, "Unhandled operation!", node_->kind().toQualString());
+    }
+  }
+  DCHECK_EQ(stack.size(), node_->outputs().size());
+  for (auto i = 0; i < node_->outputs().size(); i++) {
+    workspace[node_->outputs()[i]] = stack[i];
   }
 }